Refine scale armor example
